pyproject.toml
[project]
name = "travel-planner-agent-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.23",
    "langchain-deepseek>=0.1.3",
    "langchain-experimental>=0.3.4",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.3",
    "loguru>=0.7.3",
    "lxml>=5.4.0",
    "markdown2>=2.5.3",
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "tabulate>=0.9.0",
    "serpapi>=0.1.5"
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
]

